requirements.txt
# Core ML
torch>=2.0.0
transformers>=5.1.0
peft>=0.12.0
trl>=0.28.0
bitsandbytes>=0.49.1
accelerate>=1.12.0
sentencepiece
protobuf

# Hugging Face
huggingface_hub>=0.24.0
datasets

# RAG
faiss-cpu>=1.7.4          # use faiss-gpu if available
sentence-transformers>=2.7.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart

# Sensor & data processing
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.4.0

# Report generation
jinja2>=3.1.4
fpdf2>=2.7.9

# Utilities
python-dotenv>=1.0.1
httpx>=0.27.0
tqdm
loguru

# Dev / testing
pytest>=8.2.0
pytest-asyncio
httpx  # for TestClient
